backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
proxmoxer
requests
django-cors-headers
openai
